criaçao da classe Players
