+++ Projeto teste redis/src/main/MesaDePoker.java
@@ -189,25 +189,4 @@ public class MesaDePoker {
         }
     }
 
-    public class Players{
-        private String nome;  // nome do jogador
-        private int stack; // numero de fichas que ele possui
-        private boolean smallBlind; // marca se ele é small blid nessa mão
-        private boolean bigBlind; // marca se ele é big blid nessa mão
-
-        public Players(String nome, int stack) {
-            this.nome = nome;
-            this.stack = stack;
-            this.smallBlind = false;
-            this.bigBlind = false;
-        }
-
-        @Override
-        public String toString() {
-            return "Players{" +
-                    "nome='" + nome + '\'' +
-                    ", stack=" + stack +
-                    '}';
-        }
-    }
 }
